+++ SilKit/source/core/vasio/VAsioPeer.hpp
@@ -38,6 +38,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 
 #include "IIoContext.hpp"
 #include "IRawByteStream.hpp"
+#include "ITimer.hpp"
 
 
 namespace SilKit {
@@ -47,6 +48,7 @@ namespace Core {
 class VAsioPeer
     : public IVAsioPeer
     , private IRawByteStreamListener
+    , private ITimerListener
 {
 public:
     // ----------------------------------------
@@ -93,6 +95,8 @@ public:
 
     void Shutdown() override;
 
+    void EnableAggregation() override;
+
 private:
     // ----------------------------------------
     // Private Methods
@@ -100,11 +104,17 @@ private:
     void WriteSomeAsync();
     void ReadSomeAsync();
     void DispatchBuffer();
+    void SendSilKitMsgInternal(std::vector<uint8_t> blob);
+    void Aggregate(const std::vector<uint8_t>& blob);
+    void Flush();
 
 private: // IRawByteStreamListener
     void OnAsyncReadSomeDone(IRawByteStream& stream, size_t bytesTransferred) override;
     void OnAsyncWriteSomeDone(IRawByteStream& stream, size_t bytesTransferred) override;
     void OnShutdown(IRawByteStream& stream) override;
+
+    // ITimerListener
+    void OnTimerExpired(ITimer& timer) override;
 
 private:
     // ----------------------------------------
@@ -130,9 +140,18 @@ private:
     std::deque<std::vector<uint8_t>> _sendingQueue;
     ConstBuffer _currentSendingBuffer;
     std::vector<uint8_t> _currentSendingBufferData;
+    std::vector<uint8_t> _aggregatedMessages;
 
     std::atomic_bool _sending{false};
     Core::ServiceDescriptor _serviceDescriptor;
+
+    bool _useAggregation{false};
+    const size_t _aggregationBufferThreshold{100 * 1000};
+
+    // we trigger a flush of aggregated messages, if too much time has passed since the last flush
+    std::unique_ptr<ITimer> _flushTimer;
+    const std::chrono::milliseconds _flushTimeout{50};
+    bool _initialTimerStarted{false};
 };
 
 // ================================================================================