+++ SilKit/source/services/rpc/RpcTestUtilities.hpp
@@ -44,8 +44,8 @@ namespace Tests {
 struct MockConnection
 {
     MockConnection(SilKit::Core::IParticipantInternal*, VSilKit::IMetricsManager*,
-                   SilKit::Config::ParticipantConfiguration /*config*/,
-                   std::string /*participantName*/, SilKit::Core::ParticipantId /*participantId*/,
+                   SilKit::Config::ParticipantConfiguration /*config*/, std::string /*participantName*/,
+                   SilKit::Core::ParticipantId /*participantId*/,
                    SilKit::Services::Orchestration::ITimeProvider* /*timeProvider*/, SilKit::Core::ProtocolVersion)
     {
     }
@@ -114,6 +114,7 @@ struct MockConnection
     void FlushSendBuffers() {}
     void ExecuteDeferred(std::function<void()> /*callback*/) {}
     void NotifyShutdown() {}
+    void EnableAggregation() {}
 
     void RegisterMessageReceiver(
         std::function<void(SilKit::Core::IVAsioPeer* /*peer*/, SilKit::Core::ParticipantAnnouncement)> /*callback*/)